requirements.txt
asn1crypto==1.4.0
astroid==2.4.1
azure-common==1.1.26
azure-core==1.9.0
azure-storage-blob==12.6.0
boto3==1.15.18
botocore==1.18.18
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
coverage==5.2.1
cryptography==3.2.1
deprecation==2.1.0
docutils==0.15.2
dodgy==0.2.1
flake8==3.8.3
flake8-polyfill==1.0.2
gitdb==4.0.5
GitPython==3.1.11
google-api-core==1.23.0
google-auth==1.23.0
google-cloud-core==1.4.3
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
idna==2.10
isodate==0.6.0
isort==4.3.21
JayDeBeApi==1.2.3
jmespath==0.10.0
JPype1==1.1.2
lazy-object-proxy==1.4.3
logitol==0.1.0.0
mccabe==0.6.1
msrest==0.6.19
mysql-connector-python==8.0.22
nose2==0.9.2
numpy==1.19.3
oauthlib==3.1.0
oscrypto==1.2.1
packaging==20.4
pandas==1.1.4
pep8-naming==0.10.0
prospector==1.3.0
protobuf==3.14.0
providah==0.1.15.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pycryptodomex==3.9.9
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.5.2
pylint-celery==0.3
pylint-django==2.0.15
pylint-flask==0.6
pylint-plugin-utils==0.6
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-json-logger==2.0.1
pytz==2020.4
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
requirements-detector==0.7
rsa==4.6
s3transfer==0.3.3
setoptconf==0.2.0
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
snowflake-connector-python==2.3.6
SQLAlchemy==1.3.20
toml==0.10.1
urllib3==1.25.11
wrapt==1.12.1
